Feat: STEP 2 InputRange 열거형 / BodyCondition 구조체 작성 STEP 1에서 복사
